requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
PyMySQL==1.0.2
redis==3.5.3
requests==2.25.1
SQLAlchemy==1.4.19
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
